feat: Update ComfyUI Installation Script
This commit updates the ComfyUI installation script for the Linux environment. Notable change includes switching the repository URL for ComfyUI-Manager from https://github.com/ltdrdata/ComfyUI-Manager to https://github.com/builtwithai/cuim, presumably to pull from a different or updated version of the manager.

Another addition to the script is the execution of run_cpu.sh at the end, which is done via the bash -x command. This change ensures that upon setting up and activating the environment, the main application is run on a CPU. Thus making the end-to-end setup process more autonomous and quicker 🚀.

The permissions for run_cpu.sh have also been updated to be executable. This will eliminate potential permission issues from the mix. 💼

Remember: This change only affects the ComfyUI installation script for Linux! Other platforms are not impacted.