feat: Update repository URL 🔄
This commit modifies the repository URL from "https://github.com/ltdrdata/ComfyUI-Manager" to "https://github.com/builtwithai/cuim" throughout the application.

This includes urls (1) in the README.md, such as the install instructions 📝, (2) in the different JSON files, such as the custom-node-list.json 📄, (3) in the Python files, like __init__.py 🐍, and (4) in installation scripts files like install-comfyui-venv-win.bat 📑.

Take note that this change is necessary for the application to function correctly, as it needs to point to the correct and up-to-date repository to clone the necessary files for installation and runtime procedures.

Thus, the change is comprehensive - affecting several files across different parts of the codebase to ensure consistency and keep everything synced with the updated repository path. 🔄🔄🔄